+++ tests/unit/test_charm.py
@@ -12,18 +12,13 @@ from ops.testing import Container, Context, PeerRelation, Relation, Secret, Stat
 from constants import JOB_OCI_IMAGE, KYUUBI_CONTAINER_NAME
 from core.domain import Status
 
+from .helpers import (
+    parse_spark_properties,
+)
+
 logger = logging.getLogger(__name__)
 SPARK_PROPERTIES = "/etc/spark8t/conf/spark-defaults.conf"
 KYUUBI_PROPERTIES = "/opt/kyuubi/conf/kyuubi-defaults.conf"
-
-
-def parse_spark_properties(tmp_path: Path) -> dict[str, str]:
-    """Parse and return spark properties from the conf file in the container."""
-    file_path = tmp_path / Path(SPARK_PROPERTIES).relative_to("/etc")
-    with file_path.open("r") as fid:
-        return dict(
-            row.rsplit("=", maxsplit=1) for line in fid.readlines() if (row := line.strip())
-        )
 
 
 def parse_kyuubi_configurations(tmp_path: Path) -> dict[str, str]:
