+++ tests/unit/test_config.py
@@ -9,11 +9,15 @@ from unittest.mock import patch
 
 import pytest
 import yaml
-from ops.testing import Container, Context, PeerRelation, Relation, Secret, State
+from ops.testing import Container, Context, PeerRelation, Relation, State
 
 from charm import KyuubiCharm
 from constants import KYUUBI_CONTAINER_NAME
 from core.domain import Status
+
+from .helpers import (
+    parse_spark_properties,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -21,17 +25,6 @@ logger = logging.getLogger(__name__)
 CONFIG = yaml.safe_load(Path("./config.yaml").read_text())
 ACTIONS = yaml.safe_load(Path("./actions.yaml").read_text())
 METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
-
-SPARK_PROPERTIES = "/etc/spark8t/conf/spark-defaults.conf"
-
-
-def parse_spark_properties(tmp_path: Path) -> dict[str, str]:
-    """Parse and return spark properties from the conf file in the container."""
-    file_path = tmp_path / Path(SPARK_PROPERTIES).relative_to("/etc")
-    with file_path.open("r") as fid:
-        return dict(
-            row.rsplit("=", maxsplit=1) for line in fid.readlines() if (row := line.strip())
-        )
 
 
 @pytest.fixture()
@@ -77,6 +70,7 @@ def charm_configuration():
 @patch(
     "config.spark.SparkConfig._sa_conf", return_value={"spark.hadoop.fs.s3a.endpoint": "foo.bar"}
 )
+@pytest.mark.parametrize("profile", ["testing", "production"])
 def test_profile_config_option(
     mock_sa_conf,
     mock_s3_configured,
@@ -94,46 +88,22 @@ def test_profile_config_option(
     auth_db_relation: Relation,
     kyuubi_peers_relation: PeerRelation,
     tmp_path,
+    profile,
 ) -> None:
-    """Test when the admin password supplied via system-users config option is valid."""
-    system_users_secret = Secret(tracked_content={"admin": "password"})
+    """Test profile config option."""
     state = State(
         relations=[spark_service_account_relation, auth_db_relation, kyuubi_peers_relation],
         containers=[kyuubi_container],
-        config={"system-users": system_users_secret.id, "profile": "testing"},
-        secrets=[system_users_secret],
+        config={"profile": profile},
         leader=True,
     )
     out = kyuubi_context.run(kyuubi_context.on.config_changed(), state)
     assert out.unit_status == Status.ACTIVE.value
 
-    assert mock_set_password.called
-    _, kwargs = mock_set_password.call_args
-    assert kwargs["username"] == "admin"
-    assert kwargs["password"] == "password"
-
-    peer_app_secret = [
-        secret for secret in out.secrets if secret.label == "kyuubi-peers.kyuubi-k8s.app"
-    ]
-    assert len(peer_app_secret) > 0
-    peer_app_secret_content = peer_app_secret[0].latest_content
-    assert peer_app_secret_content is not None
-    assert peer_app_secret_content["admin-password"] == "password"
-
     spark_properties = parse_spark_properties(tmp_path)
     logger.info(f"profile: testing => spark_properties: {spark_properties}")
-    assert "spark.kubernetes.executor.request.cores" in spark_properties
-
-    state_in = State(
-        relations=[spark_service_account_relation, auth_db_relation, kyuubi_peers_relation],
-        containers=[kyuubi_container],
-        config={"profile": "production"},
-        secrets=[system_users_secret],
-        leader=True,
-    )
-    out = kyuubi_context.run(kyuubi_context.on.config_changed(), state_in)
-    assert out.unit_status == Status.ACTIVE.value
-    spark_properties = parse_spark_properties(tmp_path)
-    logger.info(f"profile: production => spark_properties: {spark_properties}")
-
-    assert "spark.kubernetes.executor.request.cores" not in spark_properties
+    if profile == "testing":
+        assert "spark.kubernetes.executor.request.cores" in spark_properties
+        assert spark_properties["spark.kubernetes.executor.request.cores"] == "0.1"
+    else:
+        assert "spark.kubernetes.executor.request.cores" not in spark_properties
